Cronet upstreaming and Workmanager update (#109)
Co-authored-by: mnbogner <mnbogner@eds.org>

This has a junky history, so I'm squash-merging it. This PR has an (I hope) fair summary of the changes (https://github.com/greatfire/envoy/pull/109). This is a major revamp of Envoy to move off of deprecated APIs (IntentService) and make the Interceptor more flexible. for example, we can use SOCKS proxies though Cronet or directly though OkHttp. This hopefully leaves things more understandable and maintainable